Massive changes
* Code is much more organized.
* Split stuff into their own namespaces
* Going to have a "raw" and "wrapped" api, you can disable it by disabling default features or just access with vm::raw::lua_* instead of vm::lua_*
* Created wrappers & raw bindings for luau_load & luaL_checkoption
All of this is unstable and will most likely be changed. (Biggest example is that raw/mod.rs will have all of the lua type definitions moved to top, probably.)
Very much WIP. Pushing these as there's too many changes now and it's better to keep track of this with smaller commits.